Reorganized Description block test cases + new tests for DTO factories
