+++ tests/Description/LanguageDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Description;
+
+use Grixu\SociusModels\Description\Factories\LanguageDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class LanguageDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = LanguageDataFactory::class;
+}
