+++ tests/Description/LanguageMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Description;
 
 use CreateLanguagesTable;
 use Grixu\SociusModels\Tests\Helpers\MigrationTest;
@@ -13,7 +13,7 @@ class LanguageMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/description/2020_10_05_085814_create_languages_table.php';
+        require_once __DIR__ . '/../../migrations/description/2020_10_05_085814_create_languages_table.php';
         $this->obj = new CreateLanguagesTable();
     }
 }
