+++ tests/Description/ProductDescriptionDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Description;
+
+use Grixu\SociusModels\Description\Factories\ProductDescriptionDataFactory;
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+
+class ProductDescriptionDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = ProductDescriptionDataFactory::class;
+}
