+++ tests/Description/ProductDescriptionMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Description;
 
 use CreateLanguagesTable;
 use CreateProductDescriptionsTable;
@@ -16,13 +16,13 @@ class ProductDescriptionMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/description/2020_10_05_100104_create_product_descriptions_table.php';
+        require_once __DIR__ . '/../../migrations/description/2020_10_05_100104_create_product_descriptions_table.php';
         $this->obj = new CreateProductDescriptionsTable();
     }
 
     public function product_relationship(): void
     {
-        require_once __DIR__ . '/../migrations/product/2020_09_25_081823_create_products_table.php';
+        require_once __DIR__ . '/../../migrations/product/2020_09_25_081823_create_products_table.php';
         (new CreateProductsTable())->up();
 
         $this->up();
@@ -33,7 +33,7 @@ class ProductDescriptionMigrationTest extends MigrationTest
     /** @test */
     public function language_relationship(): void
     {
-        require_once __DIR__ . '/../migrations/description/2020_10_05_085814_create_languages_table.php';
+        require_once __DIR__ . '/../../migrations/description/2020_10_05_085814_create_languages_table.php';
         (new CreateLanguagesTable())->up();
 
         $this->up();
